@@ -82,20 +82,20 @@
         "deploy:ci": "ng deploy --no-build --message=\"Release $npm_package_name (v$npm_package_version) on gh-pages\""
     },
     "dependencies": {
-        "@angular/animations": "^17.2.4",
-        "@angular/common": "^17.2.4",
-        "@angular/compiler": "^17.2.4",
-        "@angular/core": "^17.2.4",
-        "@angular/forms": "^17.2.4",
-        "@angular/localize": "^17.2.4",
-        "@angular/platform-browser": "^17.2.4",
-        "@angular/platform-browser-dynamic": "^17.2.4",
-        "@angular/platform-server": "^17.2.4",
-        "@angular/router": "^17.2.4",
-        "@codemirror/legacy-modes": "^6.3.3",
+        "@angular/animations": "^17.3.5",
+        "@angular/common": "^17.3.5",
+        "@angular/compiler": "^17.3.5",
+        "@angular/core": "^17.3.5",
+        "@angular/forms": "^17.3.5",
+        "@angular/localize": "^17.3.5",
+        "@angular/platform-browser": "^17.3.5",
+        "@angular/platform-browser-dynamic": "^17.3.5",
+        "@angular/platform-server": "^17.3.5",
+        "@angular/router": "^17.3.5",
+        "@codemirror/legacy-modes": "^6.4.0",
         "@fortawesome/angular-fontawesome": "^0.14.1",
-        "@fortawesome/fontawesome-svg-core": "^6.5.1",
-        "@fortawesome/free-solid-svg-icons": "^6.5.1",
+        "@fortawesome/fontawesome-svg-core": "^6.5.2",
+        "@fortawesome/free-solid-svg-icons": "^6.5.2",
         "@kolkov/ngx-gallery": "2.0.1",
         "@ng-bootstrap/ng-bootstrap": "^16.0.0",
         "@popperjs/core": "^2.11.8",
