@@ -1,15 +1,15 @@
 // THIS IS AN AUTO-GENERATED FILE. DO NOT CHANGE IT MANUALLY!
-// Generated last time on Thu, Mar  7, 2024  2:49:33 PM
+// Generated last time on Tue, Apr 23, 2024  3:51:46 PM
 
 /**
  * The latest version of the AWG App
  */
-export const appVersion = '0.11.7';
+export const appVersion = '0.12.0';
 
 /**
  * The release date of the latest version of the AWG App
  */
-export const appVersionReleaseDate = '07. März 2024';
+export const appVersionReleaseDate = '23. April 2024';
 
 /**
  * The URL of the AWG App
