@@ -1,8 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { Component, DebugElement } from '@angular/core';
 import { ComponentFixture, fakeAsync, TestBed, waitForAsync } from '@angular/core/testing';
-import { Router } from '@angular/router';
-import { RouterTestingModule } from '@angular/router/testing';
+import { Router, RouterModule } from '@angular/router';
 
 import { EmptyError, lastValueFrom, Observable, of as observableOf, throwError as observableThrowError } from 'rxjs';
 import Spy = jasmine.Spy;
@@ -14,6 +13,8 @@ import { clickAndAwaitChanges } from '@testing/click-helper';
 import { detectChangesOnPush } from '@testing/detect-changes-on-push-helper';
 import {
     expectSpyCall,
+    expectToBe,
+    expectToEqual,
     getAndExpectDebugElementByCss,
     getAndExpectDebugElementByDirective,
 } from '@testing/expect-helper';
@@ -52,7 +53,8 @@ describe('IntroComponent (DONE)', () => {
     let expectedNextComplexId: string;
     let expectedEditionIntroData: IntroList;
     let expectedEditionIntroEmptyData: IntroList;
-    let expectedFragment: string;
+    let expectedIntroFragment: string;
+    let expectedReportFragment: string;
     let expectedModalSnippet: string;
     let expectedSvgSheet: EditionSvgSheet;
     let expectedNextSvgSheet: EditionSvgSheet;
@@ -61,6 +63,8 @@ describe('IntroComponent (DONE)', () => {
     let editionDataServiceGetEditionIntroDataSpy: Spy;
     let getEditionIntroDataSpy: Spy;
     let getEditionComplexSpy: Spy;
+    let navigateToFragmentSpy: Spy;
+    let navigateToIntroFragmentSpy: Spy;
     let navigateToReportFragmentSpy: Spy;
     let navigationSpy: Spy;
     let modalOpenSpy: Spy;
@@ -81,7 +85,7 @@ describe('IntroComponent (DONE)', () => {
         };
 
         TestBed.configureTestingModule({
-            imports: [NgbModalModule, RouterTestingModule],
+            imports: [NgbModalModule, RouterModule],
             declarations: [CompileHtmlComponent, EditionIntroComponent, ModalStubComponent, RouterLinkStubDirective],
             providers: [
                 { provide: EditionDataService, useValue: mockEditionDataService },
@@ -105,7 +109,8 @@ describe('IntroComponent (DONE)', () => {
         expectedEditionComplexBaseRoute = '/edition/complex/op12/';
         expectedComplexId = 'testComplex1';
         expectedNextComplexId = 'testComplex2';
-        expectedFragment = 'source_A';
+        expectedIntroFragment = 'footnote-80';
+        expectedReportFragment = 'source_A';
         expectedModalSnippet = JSON.parse(JSON.stringify(mockEditionData.mockModalSnippet));
         expectedEditionIntroData = JSON.parse(JSON.stringify(mockEditionData.mockIntroData));
         expectedEditionIntroEmptyData = JSON.parse(JSON.stringify(mockEditionData.mockIntroEmptyData));
@@ -120,7 +125,9 @@ describe('IntroComponent (DONE)', () => {
             observableOf(expectedEditionComplex)
         );
         getEditionIntroDataSpy = spyOn(component, 'getEditionIntroData').and.callThrough();
+        navigateToIntroFragmentSpy = spyOn(component, 'navigateToIntroFragment').and.callThrough();
         navigateToReportFragmentSpy = spyOn(component, 'navigateToReportFragment').and.callThrough();
+        navigateToFragmentSpy = spyOn(component as any, '_navigateToFragment').and.callThrough();
         navigationSpy = mockRouter.navigate as jasmine.Spy;
         modalOpenSpy = spyOn(component.modal, 'open').and.callThrough();
         componentOpenModalSpy = spyOn(component, 'openModal').and.callThrough();
@@ -145,10 +152,7 @@ describe('IntroComponent (DONE)', () => {
         });
 
         it('... should have `editionRouteConstants`', () => {
-            expect(component.editionRouteConstants).toBeDefined();
-            expect(component.editionRouteConstants)
-                .withContext(`should be ${expectedEditionRouteConstants}`)
-                .toBe(expectedEditionRouteConstants);
+            expectToEqual(component.editionRouteConstants, expectedEditionRouteConstants);
         });
 
         describe('VIEW', () => {
@@ -189,17 +193,12 @@ describe('IntroComponent (DONE)', () => {
         });
 
         it('... should have editionComplex', () => {
-            expect(component.editionComplex).toBeDefined();
-            expect(component.editionComplex)
-                .withContext(`should equal ${expectedEditionComplex}`)
-                .toEqual(expectedEditionComplex);
+            expectToEqual(component.editionComplex, expectedEditionComplex);
         });
 
         it('... should have editionIntroData$', waitForAsync(() => {
             expectAsync(lastValueFrom(component.editionIntroData$)).toBeResolved();
-            expectAsync(lastValueFrom(component.editionIntroData$))
-                .withContext(`should be resolved to ${expectedEditionIntroData}`)
-                .toBeResolvedTo(expectedEditionIntroData);
+            expectAsync(lastValueFrom(component.editionIntroData$)).toBeResolvedTo(expectedEditionIntroData);
         }));
 
         describe('VIEW', () => {
@@ -208,31 +207,80 @@ describe('IntroComponent (DONE)', () => {
                 getAndExpectDebugElementByCss(compDe, 'div.awg-intro-view', 1, 1);
             });
 
-            it('... should contain as many p elements in main div as paragraphs in intro data', () => {
+            it('... should contain as many intro paragraph elements in div.awg-intro-view as content items in intro data', () => {
                 // Div debug element
                 const divDes = getAndExpectDebugElementByCss(compDe, 'div.awg-intro-view', 1, 1);
 
                 getAndExpectDebugElementByCss(
                     divDes[0],
-                    'p.awg-intro-entry',
+                    'p.awg-intro-paragraph',
                     expectedEditionIntroData.intro[0].content.length,
                     expectedEditionIntroData.intro[0].content.length
                 );
             });
 
-            it('... should have one anchor in first paragraph, and two in the second one', () => {
+            it('... should have one anchor in first paragraph, and 3 in the second one', () => {
                 // Div debug element
                 const divDes = getAndExpectDebugElementByCss(compDe, 'div.awg-intro-view', 1, 1);
 
                 const pDes = getAndExpectDebugElementByCss(
                     divDes[0],
-                    'p.awg-intro-entry',
+                    'p.awg-intro-paragraph',
                     expectedEditionIntroData.intro[0].content.length,
                     expectedEditionIntroData.intro[0].content.length
                 );
 
                 getAndExpectDebugElementByCss(pDes[0], 'a', 1, 1);
-                getAndExpectDebugElementByCss(pDes[1], 'a', 2, 2);
+                getAndExpectDebugElementByCss(pDes[1], 'a', 3, 3);
+            });
+
+            it('... should contain as many footnote paragraphs in div.awg-intro-view as footnote items in intro data', () => {
+                // Div debug element
+                const divDes = getAndExpectDebugElementByCss(compDe, 'div.awg-intro-view', 1, 1);
+
+                getAndExpectDebugElementByCss(
+                    divDes[0],
+                    'p.awg-intro-footnote',
+                    expectedEditionIntroData.intro[0].footnotes.length,
+                    expectedEditionIntroData.intro[0].footnotes.length
+                );
+            });
+
+            it('... should have each footnote paragraph embedded in small element', () => {
+                // Div debug element
+                const divDes = getAndExpectDebugElementByCss(compDe, 'div.awg-intro-view', 1, 1);
+
+                getAndExpectDebugElementByCss(
+                    divDes[0],
+                    'small p.awg-intro-footnote',
+                    expectedEditionIntroData.intro[0].footnotes.length,
+                    expectedEditionIntroData.intro[0].footnotes.length
+                );
+            });
+
+            it('... should have one anchor in footnote paragraph', () => {
+                // Div debug element
+                const divDes = getAndExpectDebugElementByCss(compDe, 'div.awg-intro-view', 1, 1);
+
+                const pDes = getAndExpectDebugElementByCss(
+                    divDes[0],
+                    'p.awg-intro-footnote',
+                    expectedEditionIntroData.intro[0].footnotes.length,
+                    expectedEditionIntroData.intro[0].footnotes.length
+                );
+
+                getAndExpectDebugElementByCss(pDes[0], 'a', 1, 1);
+            });
+
+            it('... should contain one horizontal line and header for footnotes ', () => {
+                // Div debug element
+                const divDes = getAndExpectDebugElementByCss(compDe, 'div.awg-intro-view', 1, 1);
+
+                getAndExpectDebugElementByCss(divDes[0], 'hr', 1, 1);
+                const headerDe = getAndExpectDebugElementByCss(divDes[0], 'h5', 1, 1);
+                const headerEl = headerDe[0].nativeElement;
+
+                expectToBe(headerEl.textContent.trim(), 'Anmerkungen');
             });
 
             it('... should contain a placeholder if content of intro data is empty', waitForAsync(() => {
@@ -258,10 +306,7 @@ describe('IntroComponent (DONE)', () => {
                 const introPlaceholder = `[Die Einleitung zum Editionskomplex ${expectedEditionComplex.complexId.full} erscheint im Zusammenhang der vollständigen Edition von ${expectedEditionComplex.complexId.short} in ${expectedEditionRouteConstants.EDITION.short} ${expectedEditionComplex.series.short}/${expectedEditionComplex.section.short}.]`;
                 const strippedIntroPlaceholder = introPlaceholder.replace(/<em>/g, '').replace(/<\/em>/g, '');
 
-                expect(pEl.textContent).toBeTruthy();
-                expect(pEl.textContent.trim())
-                    .withContext(`should be ${strippedIntroPlaceholder}`)
-                    .toEqual(strippedIntroPlaceholder);
+                expectToBe(pEl.textContent.trim(), strippedIntroPlaceholder);
             }));
         });
 
@@ -277,10 +322,7 @@ describe('IntroComponent (DONE)', () => {
             it('... should have got `editionComplex` from editionService', () => {
                 expectSpyCall(getEditionComplexSpy, 1);
 
-                expect(component.editionComplex).toBeTruthy();
-                expect(component.editionComplex)
-                    .withContext(`should equal ${expectedEditionComplex}`)
-                    .toEqual(expectedEditionComplex);
+                expectToEqual(component.editionComplex, expectedEditionComplex);
             });
 
             it('... should have got editionIntroData from editionDataService', () => {
@@ -303,6 +345,42 @@ describe('IntroComponent (DONE)', () => {
             }));
         });
 
+        describe('#navigateToIntroFragment()', () => {
+            it('... should have a method `navigateToIntroFragment`', () => {
+                expect(component.navigateToIntroFragment).toBeDefined();
+            });
+
+            it('... should trigger on click', fakeAsync(() => {
+                const divDes = getAndExpectDebugElementByCss(compDe, 'div.awg-intro-view', 1, 1);
+
+                // Find paragraphs
+                const pDes = getAndExpectDebugElementByCss(
+                    divDes[0],
+                    'p.awg-intro-paragraph',
+                    expectedEditionIntroData.intro[0].content.length,
+                    expectedEditionIntroData.intro[0].content.length
+                );
+
+                // Find anchor in second paragraph
+                const anchorDes = getAndExpectDebugElementByCss(pDes[1], 'a', 3, 3);
+
+                // CLick on anchor (with navigateToIntroFragment call)
+                clickAndAwaitChanges(anchorDes[2], fixture);
+
+                expectSpyCall(navigateToIntroFragmentSpy, 1, expectedIntroFragment);
+            }));
+
+            it('... should trigger `_navigateToFragment()` method with correct parameters', () => {
+                component.navigateToIntroFragment(expectedIntroFragment);
+                fixture.detectChanges();
+
+                expectSpyCall(navigateToFragmentSpy, 1, [
+                    expectedEditionRouteConstants.EDITION_INTRO.route,
+                    expectedIntroFragment,
+                ]);
+            });
+        });
+
         describe('#navigateToReportFragment()', () => {
             it('... should have a method `navigateToReportFragment`', () => {
                 expect(component.navigateToReportFragment).toBeDefined();
@@ -314,7 +392,7 @@ describe('IntroComponent (DONE)', () => {
                 // Find paragraphs
                 const pDes = getAndExpectDebugElementByCss(
                     divDes[0],
-                    'p',
+                    'p.awg-intro-paragraph',
                     expectedEditionIntroData.intro[0].content.length,
                     expectedEditionIntroData.intro[0].content.length
                 );
@@ -325,52 +403,16 @@ describe('IntroComponent (DONE)', () => {
                 // CLick on anchor (with navigateToReportFragment call)
                 clickAndAwaitChanges(anchorDes[0], fixture);
 
-                expectSpyCall(navigateToReportFragmentSpy, 1, expectedFragment);
+                expectSpyCall(navigateToReportFragmentSpy, 1, expectedReportFragment);
             }));
 
-            it('... should navigate to fragment if given', () => {
-                component.navigateToReportFragment(expectedFragment);
+            it('... should trigger `_navigateToFragment()` method with correct parameters', () => {
+                component.navigateToReportFragment(expectedReportFragment);
                 fixture.detectChanges();
 
-                const qp = { fragment: expectedFragment };
-                expectSpyCall(navigateToReportFragmentSpy, 1, expectedFragment);
-                expectSpyCall(navigationSpy, 1, [
-                    [expectedEditionComplexBaseRoute, expectedEditionRouteConstants.EDITION_REPORT.route],
-                    qp,
-                ]);
-
-                const otherFragment = 'otherFragment';
-                qp.fragment = otherFragment;
-                component.navigateToReportFragment(otherFragment);
-                fixture.detectChanges();
-
-                expectSpyCall(navigateToReportFragmentSpy, 2, otherFragment);
-                expectSpyCall(navigationSpy, 2, [
-                    [expectedEditionComplexBaseRoute, expectedEditionRouteConstants.EDITION_REPORT.route],
-                    qp,
-                ]);
-            });
-
-            it('... should navigate without fragment if none is given', () => {
-                component.navigateToReportFragment(expectedFragment);
-                fixture.detectChanges();
-
-                const qp = { fragment: expectedFragment };
-                expectSpyCall(navigateToReportFragmentSpy, 1, expectedFragment);
-                expectSpyCall(navigationSpy, 1, [
-                    [expectedEditionComplexBaseRoute, expectedEditionRouteConstants.EDITION_REPORT.route],
-                    qp,
-                ]);
-
-                const noFragment = '';
-                qp.fragment = noFragment;
-                component.navigateToReportFragment(noFragment);
-                fixture.detectChanges();
-
-                expectSpyCall(navigateToReportFragmentSpy, 2, '');
-                expectSpyCall(navigationSpy, 2, [
-                    [expectedEditionComplexBaseRoute, expectedEditionRouteConstants.EDITION_REPORT.route],
-                    qp,
+                expectSpyCall(navigateToFragmentSpy, 1, [
+                    expectedEditionRouteConstants.EDITION_REPORT.route,
+                    expectedReportFragment,
                 ]);
             });
         });
@@ -386,13 +428,13 @@ describe('IntroComponent (DONE)', () => {
                 // Find paragraphs
                 const pDes = getAndExpectDebugElementByCss(
                     divDes[0],
-                    'p',
+                    'p.awg-intro-paragraph',
                     expectedEditionIntroData.intro[0].content.length,
                     expectedEditionIntroData.intro[0].content.length
                 );
 
                 // Find anchors in second paragraph
-                const anchorDes = getAndExpectDebugElementByCss(pDes[1], 'a', 2, 2);
+                const anchorDes = getAndExpectDebugElementByCss(pDes[1], 'a', 3, 3);
 
                 // CLick on anchor (with openModal call)
                 clickAndAwaitChanges(anchorDes[1], fixture);
@@ -450,13 +492,13 @@ describe('IntroComponent (DONE)', () => {
                 // Find paragraphs
                 const pDes = getAndExpectDebugElementByCss(
                     divDes[0],
-                    'p',
+                    'p.awg-intro-paragraph',
                     expectedEditionIntroData.intro[0].content.length,
                     expectedEditionIntroData.intro[0].content.length
                 );
 
                 // Find anchors in second paragraph
-                const anchorDes = getAndExpectDebugElementByCss(pDes[1], 'a', 2, 2);
+                const anchorDes = getAndExpectDebugElementByCss(pDes[1], 'a', 3, 3);
 
                 // CLick on anchor (with selectSvgSheet call)
                 clickAndAwaitChanges(anchorDes[0], fixture);
@@ -464,7 +506,7 @@ describe('IntroComponent (DONE)', () => {
                 expectSpyCall(selectSvgSheetSpy, 1, [expectedComplexId, expectedSvgSheet.id]);
             }));
 
-            it('... should navigate within same complex to id if given', () => {
+            it('... should navigate within same complex to sheet id if given', () => {
                 expectedComplexId = expectedEditionComplex.complexId.route.replace('/', '');
                 const expectedSheetRoute = [
                     expectedEditionComplexBaseRoute,
@@ -490,7 +532,7 @@ describe('IntroComponent (DONE)', () => {
                 expectSpyCall(navigationSpy, 2, [expectedSheetRoute, qp]);
             });
 
-            it('... should navigate within same complex without id if none is given', () => {
+            it('... should navigate within same complex without sheet id if none is given', () => {
                 expectedComplexId = expectedEditionComplex.complexId.route.replace('/', '');
                 const expectedSheetRoute = [
                     expectedEditionComplexBaseRoute,
@@ -507,16 +549,35 @@ describe('IntroComponent (DONE)', () => {
                 expectSpyCall(selectSvgSheetSpy, 1, [expectedComplexId, expectedSvgSheet.id]);
                 expectSpyCall(navigationSpy, 1, [expectedSheetRoute, qp]);
 
-                const noId = '';
-                qp.queryParams.id = noId;
+                const noSheetId = undefined;
+                qp.queryParams.id = '';
 
-                component.selectSvgSheet(expectedComplexId, noId);
+                component.selectSvgSheet(expectedComplexId, noSheetId);
                 fixture.detectChanges();
 
-                expectSpyCall(selectSvgSheetSpy, 2, [expectedComplexId, '']);
+                expectSpyCall(selectSvgSheetSpy, 2, [expectedComplexId, noSheetId]);
                 expectSpyCall(navigationSpy, 2, [expectedSheetRoute, qp]);
             });
-            it('... should navigate to id of antoher complex if given', () => {
+
+            it('... should navigate within same complex by default if no complex id is given', () => {
+                const noComplexId = '';
+                const expectedSheetRoute = [
+                    expectedEditionComplexBaseRoute,
+                    expectedEditionRouteConstants.EDITION_SHEETS.route,
+                ];
+                const qp = {
+                    queryParams: { id: expectedSvgSheet.id },
+                    queryParamsHandling: '',
+                };
+
+                component.selectSvgSheet(noComplexId, expectedSvgSheet.id);
+                fixture.detectChanges();
+
+                expectSpyCall(selectSvgSheetSpy, 1, [noComplexId, expectedSvgSheet.id]);
+                expectSpyCall(navigationSpy, 1, [expectedSheetRoute, qp]);
+            });
+
+            it('... should navigate to another complex with sheet id if given', () => {
                 expectedComplexId = expectedEditionComplex.complexId.route.replace('/', '');
                 const expectedSheetRoute = [
                     expectedEditionComplexBaseRoute,
@@ -546,7 +607,7 @@ describe('IntroComponent (DONE)', () => {
                 expectSpyCall(navigationSpy, 2, [expectedNextSheetRoute, qp]);
             });
 
-            it('... should navigate to another complex without id if none is given', () => {
+            it('... should navigate to another complex without sheet id if none is given', () => {
                 expectedComplexId = expectedEditionComplex.complexId.route.replace('/', '');
                 const expectedSheetRoute = [
                     expectedEditionComplexBaseRoute,
@@ -563,18 +624,71 @@ describe('IntroComponent (DONE)', () => {
                 expectSpyCall(selectSvgSheetSpy, 1, [expectedComplexId, expectedSvgSheet.id]);
                 expectSpyCall(navigationSpy, 1, [expectedSheetRoute, qp]);
 
-                const noId = '';
-                qp.queryParams.id = noId;
+                const noSheetId = undefined;
+                qp.queryParams.id = '';
                 const expectedNextSheetRoute = [
                     `/edition/complex/${expectedNextComplexId}/`,
                     expectedEditionRouteConstants.EDITION_SHEETS.route,
                 ];
 
-                component.selectSvgSheet(expectedNextComplexId, noId);
+                component.selectSvgSheet(expectedNextComplexId, noSheetId);
                 fixture.detectChanges();
 
-                expectSpyCall(selectSvgSheetSpy, 2, [expectedNextComplexId, '']);
+                expectSpyCall(selectSvgSheetSpy, 2, [expectedNextComplexId, noSheetId]);
                 expectSpyCall(navigationSpy, 2, [expectedNextSheetRoute, qp]);
+            });
+        });
+
+        describe('#_navigateToFragment()', () => {
+            it('... should have a method `_navigateToFragment`', () => {
+                expect((component as any)._navigateToFragment).toBeDefined();
+            });
+
+            it('... should navigate to a given intro route with the given intro fragment', () => {
+                (component as any)._navigateToFragment(
+                    expectedEditionRouteConstants.EDITION_INTRO.route,
+                    expectedIntroFragment
+                );
+                fixture.detectChanges();
+
+                const qp = { fragment: expectedIntroFragment };
+                expectSpyCall(navigateToFragmentSpy, 1, [
+                    expectedEditionRouteConstants.EDITION_INTRO.route,
+                    expectedIntroFragment,
+                ]);
+                expectSpyCall(navigationSpy, 1, [
+                    [expectedEditionComplexBaseRoute, expectedEditionRouteConstants.EDITION_INTRO.route],
+                    qp,
+                ]);
+            });
+
+            it('... should navigate to a given report route with the given report fragment', () => {
+                (component as any)._navigateToFragment(
+                    expectedEditionRouteConstants.EDITION_REPORT.route,
+                    expectedReportFragment
+                );
+                fixture.detectChanges();
+
+                const qp = { fragment: expectedReportFragment };
+                expectSpyCall(navigateToFragmentSpy, 1, [
+                    expectedEditionRouteConstants.EDITION_REPORT.route,
+                    expectedReportFragment,
+                ]);
+                expectSpyCall(navigationSpy, 1, [
+                    [expectedEditionComplexBaseRoute, expectedEditionRouteConstants.EDITION_REPORT.route],
+                    qp,
+                ]);
+            });
+
+            it('... should navigate to a given route without fragmentId if none is given', () => {
+                const otherRoute = 'otherRoute';
+                const noFragment = '';
+                (component as any)._navigateToFragment(otherRoute, noFragment);
+                fixture.detectChanges();
+
+                const qp = { fragment: noFragment };
+                expectSpyCall(navigateToFragmentSpy, 1, [otherRoute, noFragment]);
+                expectSpyCall(navigationSpy, 1, [[expectedEditionComplexBaseRoute, otherRoute], qp]);
             });
         });
     });
