@@ -6,8 +6,8 @@ import {
     SourceDescriptionList,
     SourceDescriptionWritingInstruments,
     SourceDescriptionWritingMaterialDimension,
-    SourceDescriptionWritingMaterialFirmSignLocation,
     SourceDescriptionWritingMaterialFormat,
+    SourceDescriptionWritingMaterialItemLocation,
     SourceDescriptionWritingMaterialSystems,
 } from '@awg-views/edition-view/models';
 
@@ -32,6 +32,14 @@ export class SourceDescriptionComponent {
      */
     @Input()
     sourceDescriptionListData: SourceDescriptionList;
+
+    /**
+     * Output variable: navigateToReportFragment.
+     *
+     * It keeps an event emitter for a fragment id of the edition report.
+     */
+    @Output()
+    navigateToReportFragmentRequest: EventEmitter<string> = new EventEmitter();
 
     /**
      * Output variable: openModalRequest.
@@ -112,21 +120,17 @@ export class SourceDescriptionComponent {
     }
 
     /**
-     * Public method: getWritingMaterialFirmSignLocation.
+     * Public method: getWritingMaterialItemLocation.
      *
      * It retrieves the string representation of the location
-     * of the firm sign of the writing material
+     * of an item of the writing material (firm Sign or watermark)
      * provided in the source description.
      *
-     * @param {SourceDescriptionWritingMaterialFirmSignLocation} location The given location data.
+     * @param {SourceDescriptionWritingMaterialItemLocation} location The given location data.
      * @returns {string} The retrieved location string.
      */
-    getWritingMaterialFirmSignLocation(location: SourceDescriptionWritingMaterialFirmSignLocation): string {
-        if (
-            !this.utils.isNotEmptyObject(location) ||
-            !this.utils.isNotEmptyArray(location.folios) ||
-            !location.position
-        ) {
+    getWritingMaterialItemLocation(location: SourceDescriptionWritingMaterialItemLocation): string {
+        if (!this.utils.isNotEmptyObject(location) || !this.utils.isNotEmptyArray(location.folios)) {
             return '';
         }
         const foliosFormatted = location.folios.map((folio: string) =>
@@ -139,8 +143,9 @@ export class SourceDescriptionComponent {
                 : foliosFormatted[0];
 
         const info = location.info ? `${location.info} ` : '';
+        const position = location.position ? ` ${location.position}` : '';
 
-        return `${info}auf Bl. ${foliosString} ${location.position}`;
+        return `${info}auf Bl. ${foliosString}${position}`;
     }
 
     /**
@@ -182,6 +187,22 @@ export class SourceDescriptionComponent {
         ];
 
         return systemsOutput.filter(Boolean).join('');
+    }
+
+    /**
+     * Public method: navigateToReportFragment.
+     *
+     * It emits a given id of a fragment of the edition report
+     * to the {@link navigateToReportFragmentRequest}.
+     *
+     * @param {string} id The given fragment id.
+     * @returns {void} Navigates to the edition report.
+     */
+    navigateToReportFragment(id: string): void {
+        if (!id) {
+            return;
+        }
+        this.navigateToReportFragmentRequest.emit(id);
     }
 
     /**
