@@ -109,7 +109,7 @@ export class EDITION_ROUTE_CONSTANTS {
     static readonly SERIES: EditionRouteConstant = {
         route: 'series',
         short: 'Serien',
-        full: 'Serienübersicht',
+        full: 'Editionsübersicht',
     };
 
     /**
@@ -172,16 +172,16 @@ export class EDITION_ROUTE_CONSTANTS {
     static readonly SECTION_2A: EditionRouteConstant = {
         route: '2a',
         short: '2a',
-        full: 'Abteilung 2a (Klaviermusik)',
+        full: 'Abteilung 2a (Kammer- und Klaviermusik: Klaviermusik)',
     };
 
     /**
      * The route constant for section 2b.
      */
     static readonly SECTION_2B: EditionRouteConstant = {
-        route: '2a',
-        short: '2a',
-        full: 'Abteilung 2b (Kammermusik)',
+        route: '2b',
+        short: '2b',
+        full: 'Abteilung 2b (Kammer- und Klaviermusik: Kammermusik)',
     };
 
     /**
